reorganising share app specific code
