+++ backend/tracim_backend/applications/share/share_controller.py
@@ -5,32 +5,32 @@ from depot.manager import DepotManager
 from hapic.data import HapicFile
 from pyramid.config import Configurator
 
-from tracim_backend import TracimRequest
 from tracim_backend.app_models.contents import FILE_TYPE
+from tracim_backend.applications.share.schema import ContentShareInfoSchema
+from tracim_backend.applications.share.schema import ContentShareSchema
+from tracim_backend.applications.share.schema import ShareCreationBodySchema
+from tracim_backend.applications.share.schema import ShareIdPathSchema
+from tracim_backend.applications.share.schema import ShareListQuerySchema
+from tracim_backend.applications.share.schema import ShareTokenPathSchema
+from tracim_backend.applications.share.schema import ShareTokenWithFilenamePathSchema
+from tracim_backend.applications.share.schema import TracimSharePasswordHeaderSchema
+from tracim_backend.applications.share.share import ShareLib
 from tracim_backend.config import CFG
 from tracim_backend.exceptions import ContentTypeNotAllowed
 from tracim_backend.exceptions import TracimFileNotFound
 from tracim_backend.exceptions import WrongSharePassword
 from tracim_backend.extensions import hapic
-from tracim_backend.lib.share.share import ShareLib
 from tracim_backend.lib.utils.authorization import ContentTypeChecker
 from tracim_backend.lib.utils.authorization import check_right
 from tracim_backend.lib.utils.authorization import is_content_manager
 from tracim_backend.lib.utils.authorization import is_contributor
+from tracim_backend.lib.utils.request import TracimRequest
 from tracim_backend.lib.utils.utils import generate_documentation_swagger_tag
 from tracim_backend.models.content_share import ContentShare
 from tracim_backend.models.context_models import ContentShareInContext
 from tracim_backend.views.controllers import Controller
 from tracim_backend.views.core_api.schemas import NoContentSchema
 from tracim_backend.views.core_api.schemas import WorkspaceAndContentIdPathSchema
-from tracim_backend.views.share_api.schema import ContentShareInfoSchema
-from tracim_backend.views.share_api.schema import ContentShareSchema
-from tracim_backend.views.share_api.schema import ShareCreationBodySchema
-from tracim_backend.views.share_api.schema import ShareIdPathSchema
-from tracim_backend.views.share_api.schema import ShareListQuerySchema
-from tracim_backend.views.share_api.schema import ShareTokenPathSchema
-from tracim_backend.views.share_api.schema import ShareTokenWithFilenamePathSchema
-from tracim_backend.views.share_api.schema import TracimSharePasswordHeaderSchema
 from tracim_backend.views.swagger_generic_section import SWAGGER_TAG__CONTENT_ENDPOINTS
 
 SWAGGER_TAG__CONTENT_SHARE_SECTION = "Share"
