+++ backend/tracim_backend/applications/share/share_email_manager.py
