+++ backend/tracim_backend/lib/search/search_factory.py
@@ -2,7 +2,7 @@ import typing
 
 from sqlalchemy.orm import Session
 
-from tracim_backend import CFG
+from tracim_backend.config import CFG
 from tracim_backend.exceptions import NoValidSearchEngine
 from tracim_backend.models.auth import User
