+++ backend/tracim_backend/models/context_models.py
@@ -1463,7 +1463,7 @@ class ContentShareInContext(object):
     @property
     def url(self) -> str:
         # TODO - G.M - 2019-07-31 - import here to avoid recursive import.
-        from tracim_backend.lib.share.share import ShareLib
+        from tracim_backend.applications.share.share import ShareLib
 
         api = ShareLib(config=self.config, session=self.dbsession, current_user=self._user)
         return api.frontend_url(self.content_share)
@@ -1471,7 +1471,7 @@ class ContentShareInContext(object):
     @property
     def direct_url(self) -> str:
         # TODO - G.M - 2019-07-31 - import here to avoid recursive import.
-        from tracim_backend.lib.share.share import ShareLib
+        from tracim_backend.applications.share.share import ShareLib
 
         api = ShareLib(config=self.config, session=self.dbsession, current_user=self._user)
         return api.direct_api_url(self.content_share)
